adblib: Add support for "<device-prefix>:shell"
Add support for executing shell commands on devices and capturing the
command output (`stdout`).

The output is captured in "raw" (i.e. non text) format to allow
capturing output of aribtrary format. The output is captured in
"chunks" of configurable size to allow capturing output of
arbitrary size.

Extension methods are provided to capture the output as a single String
or as a list of Strings (i.e. lines) for commands that output text.

Bug: 188067646
Test: Included
Change-Id: I0416719b0c1ba3a5a90959a9c1e425a63f8cc0d1
Signed-off-by: k-x7 <97758178+k-x7@users.noreply.github.com>