adblib: implement channel based shell collector
For shell commands that return arbitrary binary data, having a raw
collector that emits `ByteBuffer` of arbitrary sizes is not convenient.

This change adds a `InputChannelShellCollector` that wraps the output
of a shell command as as `AdbInputChannel` for `stdout` and `stderr`.
The caller can then use these `AdbInputChannel` instances to
consume (as suspending function) the command output as long as the
command is active.

Given `shellCommand` returns a `Flow`, the API is a little awkward to
use with `execute`, so we also introduce a new `executeSingle`
operation on `shellCommand`, which allows processing the output
of a `ShellOutput` as a single value, keeping that value alive and
active as long as the underlying shell command is active.

This change also introduces a general purpose `AdbPipedInputChannel`
interface and implementation, which allows writing arbitrary binary
data to it (similar to `java.io.PipedInputStream`).

This reverts commit f7414e2d178ebc516cc6115c1c4fefc8d8783afb.

Reason for revert: Unit tests need to be fixed

Tests were verified by running

../bazel/bazel test --runs_per_test 5000 //tools/base/adblib:adblib.tests.test

Bug: 188067646
Test: Included
Change-Id: I9efefc0aec6a2facf0f0d50b0c347e58597d80a7
Signed-off-by: k-x7 <97758178+k-x7@users.noreply.github.com>