adblib: Use shell,v2 protocol if available
When retrieving device properties, we don't have access to the
`shellCommand` extension because the `DeviceProperties`
class is at a lower level layer.

A previous change was made to detect older `shell`
implementations that use `\r\n' for newlines (instead of '\n').

This change ensure we use the `shell,v2` protocol if supported,
to avoid the problem entirely (and be more efficient as a side
effect).

Bug: 188067646
Test: Existing
Change-Id: I901fe795f862ed5ee818572e8e7eb74e2329c0b1
Signed-off-by: k-x7 <97758178+k-x7@users.noreply.github.com>