adblib: Cleanup JdwpProcessXxx Apis
This change is mostly a refactoring of existing code, with
the intent to make the `JdwpProcessXxx` APIs more
consistent.

* Make `JdwpProcessTracker` an interface with an internal
implementation.

* Expose the flow of processes from `JdwpProcessTracker`
as a `StateFlow` instead of a regular flow.

* Use `ConnectedDevice` for both `JdwpProcess` and
`JdwpProcessTracker` (as opposed to `DeviceSelector`)
because these interfaces really only make sense for
connected devices, and their scope is tightly coupled.

* Expose a `CoroutineScope` from `JdwpProcessTracker`
that matches the lifecycle of the tracker (and the device).

* Add a couple of extension properties to `ConnectedDevice`:
`selector` and `isOnline`.

Bug: 188067646
Test: Updated
Change-Id: Ieac5a687eb5a1b0fa588f9f444016bc6da9bebca
Signed-off-by: k-x7 <97758178+k-x7@users.noreply.github.com>