adblib: Make `AdbPipedInputChannel` coroutine friendly
The existing implementation of `AdbPipedInputChannel` was using
`java.io.PipedInputStream` under the hood. This was sub-optimal as
`PipedInputStream` uses thread-blocking mechanisms to synchronize
read and writes, leading to sub-optimal use of threads.

With this change, a new implementation based on coroutine friendly
synchronization mechanisms removes this limitation and ensures
Thread are never blocked.

This change introduces a `CircularByteBuffer` class so that
`AdbPipedInputChannel` can use a single buffer for reading and
writing.

Bug: 188067646
Test: Existing + Included
Change-Id: I030b49ab2c88666435578baa5e39ca3c91ae85a7
Signed-off-by: k-x7 <97758178+k-x7@users.noreply.github.com>