adblib: Add support for "write-back" `AdbOutputChannel`
`AdbWriteBackOutputChannel` is an `AdbOutputChannel` that buffers
write operations to another `AdbOutputChannel` using an internal buffer
that is flushed by a coroutine running concurrently to the write
operations. The intent is to batch small write operations in larger
ones if the underlying `AdbOutputChannel` has high latency (e.g.
file or network channels).

The change also introduces an new `AdbBufferedOutputChannel` that
exposes a `shutdown` method, to be used by callers to ensure all
pending write operations are successfully "flushed". This cannot
be done in the `close` method, since flushing remaining data needs
to be performed in a `suspend` method.

Bug: 188067646
Test: Included
Change-Id: Ic3c75f7414a2118367ac90a3ad6463e84a890b9a
Signed-off-by: k-x7 <97758178+k-x7@users.noreply.github.com>